requirements.txt
asgiref==3.8.1
crispy-bootstrap4==2024.10
Django==5.2
django-crispy-forms==2.4
fontawesomefree==6.6.0
sqlparse==0.5.3
pillow==10.0.0